+++ nncf/onnx/quantization/quantizer_parameters.py
@@ -15,6 +15,7 @@ from typing import Optional, Tuple
 import numpy as np
 
 from nncf.quantization.fake_quantize import FakeQuantizeParameters
+from nncf.quantization.fake_quantize import calculate_scale_zero_point
 
 
 @dataclass
@@ -63,6 +64,8 @@ def convert_fq_params_to_onnx_params(
     level_low, level_high = get_level_low_level_high(tensor_type)
     narrow_range = levels == 2**num_bits - 1
     scale, zero_point = calculate_scale_zero_point(input_low, input_high, level_low, level_high, narrow_range)
+    scale = np.array(np.squeeze(scale).astype(np.float32))
+    zero_point = np.array(np.squeeze(zero_point).astype(np.int32))
     return ONNXQuantizerLayerParameters(scale, zero_point, tensor_type, axis)
 
 
@@ -75,31 +78,3 @@ def get_level_low_level_high(tensor_type: np.dtype) -> Tuple[int, int]:
     :return: Minimum level and maximum level of the quantizer.
     """
     return (0, 255) if tensor_type == np.uint8 else (-128, 127)
-
-
-def calculate_scale_zero_point(
-    input_low: np.ndarray, input_high: np.ndarray, level_low: int, level_high: int, narrow_range: bool
-) -> Tuple[np.ndarray, np.ndarray]:
-    """
-    Calculates Quantizer/Dequantizer layer scale level.
-    Returns scale and zero_point values for the quantizer.
-
-    :param input_low: The minimum limit for an input value based on collected statistics.
-    :param input_high: The maximum limit for an input value based on collected statistics.
-    :param level_low: The minimum level in the integer range to quantize.
-        The default is "0" for an unsigned range, and "-2^(bit-1)" for a signed one .
-    :param level_high: The maximum level in the integer range to quantize.
-        The default is "2^bits-1" for an unsigned range, and "2^(bit-1)-1" for a signed one.
-    :param narrow_range: True if the range of quantized values is narrowed as compared to the
-        naive case, False otherwise.
-    :return: Scale and Zero point values.
-    """
-    levels = level_high - level_low if narrow_range else level_high - level_low + 1
-    scale = np.array((input_high - input_low) / (levels - 1))
-    expected_level_low = level_low + 1 if narrow_range else level_low
-    zero_point = expected_level_low - np.round(input_low / scale)
-    zero_point = np.minimum(np.maximum(zero_point.astype(np.int32), level_low), level_high)
-    scale = np.array(np.squeeze(scale).astype(np.float32))
-    zero_point = np.array(np.squeeze(zero_point))
-
-    return scale, zero_point
