+++ nncf/openvino/graph/nncf_graph_builder.py
@@ -16,16 +16,15 @@ import openvino.runtime as ov
 from nncf.common.graph import NNCFGraph
 from nncf.common.graph.layer_attributes import Dtype
 from nncf.common.graph.operator_metatypes import OperatorMetatype
-from nncf.common.graph.operator_metatypes import UnknownMetatype
 from nncf.openvino.graph.layer_attributes import OVLayerAttributes
 from nncf.openvino.graph.layer_attributes import get_weighted_layer_attributes
 from nncf.openvino.graph.metatypes.openvino_metatypes import METATYPES_WITH_CONST_PORT_ID
-from nncf.openvino.graph.metatypes.openvino_metatypes import OV_OPERATOR_METATYPES
 from nncf.openvino.graph.metatypes.openvino_metatypes import OVConvolutionBackpropDataMetatype
 from nncf.openvino.graph.metatypes.openvino_metatypes import OVGroupConvolutionBackpropDataMetatype
 from nncf.openvino.graph.metatypes.openvino_metatypes import OVGRUSequenceMetatype
 from nncf.openvino.graph.metatypes.openvino_metatypes import OVLSTMSequenceMetatype
 from nncf.openvino.graph.metatypes.openvino_metatypes import OVMatMulMetatype
+from nncf.openvino.graph.metatypes.openvino_metatypes import get_node_metatype
 from nncf.openvino.graph.metatypes.openvino_metatypes import get_operation_const_op
 
 
@@ -79,23 +78,6 @@ class GraphConverter:
         return inputs
 
     @staticmethod
-    def _get_node_metatype(node: ov.Node) -> Type[OperatorMetatype]:
-        """
-        Determine NNCF meta type for OpenVINO node.
-
-        :param node: OpenVINO node.
-        :return: NNCF meta type which corresponds to OpenVINO node.
-        """
-        node_type = node.get_type_name()
-        metatype = OV_OPERATOR_METATYPES.get_operator_metatype_by_op_name(node_type)
-        if metatype is not UnknownMetatype:
-            if metatype.get_subtypes():
-                subtype = metatype.determine_subtype(node)
-                if subtype is not None:
-                    metatype = subtype
-        return metatype
-
-    @staticmethod
     def _add_edges_to_nncf_graph(model: ov.Model, graph: NNCFGraph) -> None:
         """
         Adds edges between NNCFNodes to the NNCFGraph.
@@ -130,7 +112,7 @@ class GraphConverter:
         :param graph: NNCFGraph.
         """
         node_type = node.get_type_name()
-        metatype = GraphConverter._get_node_metatype(node)
+        metatype = get_node_metatype(node)
         graph.add_nncf_node(node_name=node.get_friendly_name(), node_type=node_type, node_metatype=metatype)
 
     @staticmethod
@@ -159,7 +141,7 @@ class GraphConverter:
                         inference_nodes.append(inp.get_node())
 
         for node in model.get_ops():
-            metatype = GraphConverter._get_node_metatype(node)
+            metatype = get_node_metatype(node)
             # Add nodes from constant subgraphs
             node_name = node.get_friendly_name()
             if node_name not in visited:
